third_party/blink/web_tests/external/wpt/webaudio/the-audio-api/the-audiocontext-interface/audiocontext-sinkid-constructor.https-expected.txt

This is a testharness.js-based test.
PASS Setting sinkId to the empty string at construction should succeed.
FAIL Setting sinkId with a valid device identifier at construction should succeed. assert_true: the context sinkId should match the given sinkId. expected true got false
FAIL Setting sinkId with an AudioSinkOptions at construction should succeed. assert_equals: expected "object" but got "string"
FAIL Invalid sinkId arguments should throw an appropriate exception. assert_throws_js: An invalid AudioSinkOptions.type value should throw a TypeError exception. function "() => {
      const audioContext = new AudioContext({sinkId: {type: 'something_else'}});
      audioContext.close();
    }" did not throw
Harness: the test ran to completion.
